RCE-2559: Add DescriptorConfig plugin
To assist automated usage of 'rbuild launch' and 'rbuild create target' these
commands now support a --to-file option in addition to --from-file. --to-file
takes a file path where the descriptor data will be written to. These files are
suitable for then passing in as a --from-file option. Any fields that may need
editting between executions will either need to be removed from the config file,
and their values redirected into the commands, or the config file will need to
be editted prior to execution.