RCE-2620: Fix rbuild show images
rmake is no longer used to watch image builds, so the old behaviour of
'rbuild show images' no longer works. Updated the rbuild status config
file with a new options, imageJobIds, which is a list item and removed
all of the getter and setter methods linked to the old imageJobId
option. Also added methods to the RPC client to poll and show the status
of builds.