Edit targets by name and optionally type (APPENG-2863)
Targets are uniquely identified by target name and type, so this change
allows the user to specify a name, and if that is not unique they will
be asked to pick the correct target. If they know that the target name
is not unique, or want to avoid interaction, the user may provide the
type as a second command line argument.