Change the flow of 'rbuild edit user' (APPENG-2963)
Before you had to explicily select the fields you wanted to edit. This
led to the surprising behavior that 'rbuild edit user <name>' was a
no-op. Now the flow is that if no options are given, then we prompt for
the non-boolean fields: full name, email, and password.

Additionally, the behavior of the boolean flags (no-)external,
(no-)admin, and (no-)create changed. Before, you could specify
both a <flag> and its no-<flag> counterpart, and <flag> would win. Now,
those are exclusive and specifing both raises an error.