Fix interactive project/branch creation (APPENG-3046)
Some required prompts were not marked as required, and so values
could be passed in to validation functions that they could not
handle. Updated the prompt calls to mark required items as required,
and updated validation functions to check for existance of required
items. Also consolidated all of the project/branch creation dialogues
around ui.getResponse().