Enforce proper http url syntax in rbuilder coonfigs (APPENG-3106)
Currently, we do not enforce any constraints on the serverUrl or
rmakeUrl config options. However, internall calls assume that these are
valid http urls, and raise unhandled exceptions if these values are not.
Best place to enforce this is in the config file itself, so this change
introduces the CfgHttpUrl config type. Conary config parse errors are
now caught by the main handler as well, so we get nicely formatted error
messages.